+++ docs/blog/2023-03-26-debut-post.mdx
@@ -1,0 +1,29 @@
+---
+slug: debut-post
+title: Introducing Kone
+authors: [lounres]
+tags: [KMath]
+---
+
+Hi! I'm Gleb.
+I'm bachelor math student at SPbU in Saint Petersburg, Russia.
+And I introduce you Kone, a library that helps you do math experiments with more ease.
+
+# Why did I create this library?
+
+Since I was in 8th grade of school I loved to do math.
+And sometimes I unconsciously experimented by calculating a number of the first elements of different series.
+For example, the most standard problem is to found out that sum $1^k + 2^k + \dots + n^k$ for some fixed $k$ is actually
+a polynomial of $n$ (of degree $k+1$).
+And there is a lot of interesting theory around the problem.
+So sometimes I was so impatient that I calculated the first several polynomials.
+But it was such a hassle to calculate them by hand. And it still is 😉.
+The calculations were very simple to describe but very hard to calculate manually without mistakes.
+So I created a small library in Python that helped me with such a hassle.
+After that the library was growing because of new computational problems.
+And at some point it was rewritten in Kotlin.
+But it looked like hell full of non-reusable scripts of an amateur programmer
+that I rewrote it in Kotlin one more time from a scratch but now on a clean architecture.
+Also, around this time I've got to know KMath and its authors.
+So I mimicked KMath in some architectural questions.
+And using prototype of my new library I even contributed in KMath about 37K strings of code 😅.
