--- docs/docs/extra/_category_.json
@@ -1,4 +1,0 @@
-{
-  "label": "Extras",
-  "position": 1000
-}
