--- docs/docs/extra/index.mdx
@@ -1,3 +1,0 @@
-# Extras
-
-Modules described here are used in code of Kone as utilities or for formal moment, do not bear any concept, but still can be useful. That's why they're placed here. So feel free to look around this section or pass by.
