--- docs/docs/extra/mapUtil.mdx
@@ -1,9 +1,0 @@
----
-sidebar_position: 2
----
-
-# 418: I'm teapot :(
-
-Sorry, this page is currently under construction.
-
-![Just funny todo notebook](./img/todo.jpg)
