+++ docs/docs/guides/_category_.yml
@@ -1,0 +1,2 @@
+label: Guides
+position: 2
