--- docs/docs/kone-gameTheory.mdx
@@ -1,10 +1,0 @@
----
-sidebar_position: 5
----
-
-# Kone: Game Theory
-
-Sorry, this page is currently under construction.
-
-![Just funny todo notebook](./img/todo.jpg)
-
