+++ docs/docs/kone-main/_category_.yml
@@ -1,0 +1,3 @@
+label: Main modules
+position: 3
+collapsed: false
