+++ docs/docs/kone-main/algebraic.mdx
@@ -1,7 +1,8 @@
 ---
+title: Algebraic
 sidebar_position: 2
 ---
-import DependenciesTabs from '@site/src/components/DependenciesTabs';
+import DependencyTabs from '@site/src/components/DependencyTabs';
 
 # Kone: Algebraic
 
@@ -9,7 +10,7 @@ This package describes the basic concept of working with algebraic structures an
 
 ## Applying the dependencies
 
-<DependenciesTabs artifact="kone-algebraic"/>
+<DependencyTabs artifact="kone-algebraic"/>
 
 ## Concept
