+++ docs/docs/kone-main/core.mdx
@@ -1,9 +1,12 @@
 ---
+title: Core
 sidebar_position: 1
 ---
 
-# 418: I'm teapot :(
+# Kone: Core
 
+:::note
 Sorry, this page is currently under construction.
 
-![Just funny todo notebook](./img/todo.jpg)
+![Just funny todo notebook](/img/todo.jpg)
+:::
