+++ docs/docs/kone-main/gameTheory.mdx
@@ -1,0 +1,16 @@
+---
+title: Game Theory
+sidebar_position: 5
+---
+
+# Kone: Game Theory
+
+:::danger
+This module is experimental. It can be modified and even removed in the future.
+:::
+
+:::note
+Sorry, this page is currently under construction.
+
+![Just funny todo notebook](/img/todo.jpg)
+:::
