+++ docs/docs/kone-main/numberTheory.mdx
@@ -1,10 +1,13 @@
 ---
+title: Number Theory
 sidebar_position: 3
 ---
 
 # Kone: Number Theory
 
+:::note
 Sorry, this page is currently under construction.
 
-![Just funny todo notebook](./img/todo.jpg)
+![Just funny todo notebook](/img/todo.jpg)
+:::
