+++ docs/docs/kone-main/polynomial.mdx
@@ -1,10 +1,12 @@
 ---
+title: Polynomial
 sidebar_position: 4
 ---
 
 # Kone: Polynomial
 
+:::note
 Sorry, this page is currently under construction.
 
-![Just funny todo notebook](./img/todo.jpg)
-
+![Just funny todo notebook](/img/todo.jpg)
+:::
