+++ docs/docs/kone-misc/_category_.yml
@@ -1,0 +1,3 @@
+label: Misc modules
+position: 4
+collapsed: false
