+++ docs/docs/kone-misc/kone-planimetricsCalculation.mdx
@@ -1,0 +1,12 @@
+---
+title: Planimetrics Calculation
+sidebar_position: 4
+---
+
+# Kone: Misc: Planimetrics Calculation
+
+:::note
+Sorry, this page is currently under construction.
+
+![Just funny todo notebook](/img/todo.jpg)
+:::
