--- docs/docs/kone-planimetricsCalculation.mdx
@@ -1,10 +1,0 @@
----
-sidebar_position: 6
----
-
-# Kone: Planimetrics Calculation
-
-Sorry, this page is currently under construction.
-
-![Just funny todo notebook](./img/todo.jpg)
-
