+++ docs/docs/kone-util/_category_.yml
@@ -1,0 +1,3 @@
+label: Util modules
+position: 5
+collapsed: false
