+++ docs/docs/kone-util/collectionOperations.mdx
@@ -1,0 +1,12 @@
+---
+title: Collection operations
+position: 2
+---
+
+# Kone: Util: Collection operations
+
+:::note
+Sorry, this page is currently under construction.
+
+![Just funny todo notebook](/img/todo.jpg)
+:::
