+++ docs/docs/kone-util/mapOperations.mdx
@@ -1,0 +1,12 @@
+---
+title: Map operations
+position: 1
+---
+
+# Kone: Util: Map operations
+
+:::note
+Sorry, this page is currently under construction.
+
+![Just funny todo notebook](/img/todo.jpg)
+:::
