+++ docs/src/inputData.tsx
@@ -1,0 +1,2 @@
+export let koneGroup = "com.lounres.kone"
+export let koneVersion = "0.0.0-dev-1"
